MOBPLAT-380 convert junit time field to float

+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/finalizer/verdict/TestStatisticsCounter.kt
@@ -2,7 +2,7 @@ package com.avito.runner.finalizer.verdict
 
 internal interface TestStatisticsCounter {
 
-    fun overallDurationSec(): Int
+    fun overallDurationSec(): Float
 
     fun overallCount(): Int
